update prisma service
