--- backend/apps/sample-service/mikro-orm.config.ts
@@ -1,7 +1,0 @@
-import { WalletRecord } from './src/wallet/infrastructure/wallet.record';
-import { UserRecord } from './src/user/infrastructure/user.record';
-export default {
-  entities: [UserRecord, WalletRecord], // no need for `entitiesTs` this way
-  dbName: 'my-db-name',
-  type: 'mysql', // one of `mongo` | `mysql` | `mariadb` | `postgresql` | `sqlite`
-};
