+++ backend/apps/sample-service/src/app.module.ts
@@ -1,4 +1,3 @@
-import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
@@ -9,8 +8,6 @@ import {
 import { RequestContextModule } from 'nestjs-request-context';
 import { UserModule } from './user/user.module';
 import { WalletModule } from './wallet/wallet.module';
-import { UserRecord } from './user/infrastructure';
-import { WalletRecord } from './wallet/infrastructure';
 
 const interceptors = [
   {
@@ -26,16 +23,6 @@ const interceptors = [
 @Module({
   imports: [
     RequestContextModule,
-    // MikroOrmModule.forRoot({
-    //   entities: [UserRecord, WalletRecord],
-    //   dbName: 'test_db',
-    //   host: 'localhost',
-    //   user: 'root',
-    //   password: '123456',
-    //   type: 'mysql',
-    //   port: 3306,
-    //   debug: true,
-    // }),
     CqrsModule,
 
     // Modules
