+++ backend/apps/sample-service/src/main.ts
@@ -7,13 +7,10 @@ import { Logger } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 ``;
 import { AppModule } from './app.module';
-import { MikroORM } from '@mikro-orm/core';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   const globalPrefix = 'api';
-  // await app.get(MikroORM).getSchemaGenerator().ensureDatabase();
-  // await app.get(MikroORM).getSchemaGenerator().updateSchema();
   app.setGlobalPrefix(globalPrefix);
   const port = process.env.PORT || 3000;
   await app.listen(port);
