+++ backend/apps/sample-service/src/user/application/index.ts
@@ -1,2 +1,3 @@
 export * from './command-handlers';
 export * from './query-handlers';
+export * from './dtos';
