+++ backend/apps/sample-service/src/user/infrastructure/index.ts
@@ -1,2 +1,1 @@
-export * from './user.record';
 export * from './user.repository';
