+++ backend/apps/sample-service/src/user/infrastructure/user.repository.ts
@@ -1,13 +1,10 @@
-import { UserRepositoryPort } from '../domain';
+import { BaseRepository } from '@lib/common/databases';
+import { PrismaSampleService, User } from '@lib/sample-db';
+import { UserRepositoryPort, UserEntity } from '../domain';
 import { UserMapper } from '../user.mapper';
-import { UserEntity } from '../domain/user.entity';
 import { Injectable, Logger } from '@nestjs/common';
 import { EventBus } from '@nestjs/cqrs';
-import { MikroORM } from '@mikro-orm/core';
-import { UserRecord } from './user.record';
-import { EntityManager } from '@mikro-orm/mysql';
 import { None, Option, Some } from 'oxide.ts';
-import { BaseRepository } from '@lib/common/databases';
 
 /**
  *  Repository is used for retrieving/saving domain entities
@@ -15,68 +12,72 @@ import { BaseRepository } from '@lib/common/databases';
 @Injectable()
 // implements UserRepositoryPort
 export class UserRepository
-  extends BaseRepository<UserEntity, UserRecord>
+  extends BaseRepository<UserEntity, User>
   implements UserRepositoryPort
 {
   constructor(
-    protected readonly orm: MikroORM,
     protected readonly mapper: UserMapper,
     protected readonly eventBus: EventBus,
-    protected readonly em: EntityManager
+    protected readonly prisma: PrismaSampleService,
+    protected readonly logger: Logger
   ) {
-    super(orm, mapper, eventBus, new Logger('UserRepository'));
+    super(prisma, mapper, eventBus, logger);
   }
+  // transaction<T>(handler: () => Promise<T>): Promise<T> {
+  //   return this.prisma.$transaction(handler);
+  // }
 
   async insert(entity: UserEntity): Promise<boolean> {
-    const result = await this.em
-      .createQueryBuilder(UserRecord)
-      .insert({
+    const result = await this.prisma.user.create({
+      data: {
         id: entity.id,
-        createdAt: entity.createdAt,
-        updatedAt: entity.updatedAt,
+        createdAt: entity.getPropsCopy().createdAt,
+        updatedAt: entity.getPropsCopy().updatedAt,
+        email: entity.getPropsCopy().email,
+        postalCode: entity.getPropsCopy().address.postalCode,
         country: entity.getPropsCopy().address.country,
         street: entity.getPropsCopy().address.street,
-        postalCode: entity.getPropsCopy().address.postalCode,
-        email: entity.getPropsCopy().email,
         role: entity.getPropsCopy().role,
-      })
-      .execute();
+      },
+    });
     entity.publishEvents(this.logger, this.eventBus);
-    return result.affectedRows > 0;
+    return result !== null;
   }
   async findOneById(id: string): Promise<Option<UserEntity>> {
-    const result = await this.em
-      .createQueryBuilder(UserRecord, 'u')
-      .select('*')
-      .where({ id })
-      .getSingleResult();
-    return result[0] ? Some(this.mapper.toDomain(result[0])) : None;
+    const result = await this.prisma.user.findUnique({
+      where: {
+        id: id,
+      },
+    });
+    return result ? Some(this.mapper.toDomain(result)) : None;
   }
-  delete(entity: UserEntity): Promise<boolean> {
-    return this.em
-      .createQueryBuilder(UserRecord, 'u')
-      .delete()
-      .where({ id: entity.id })
-      .execute();
+  async delete(entity: UserEntity): Promise<boolean> {
+    const result = await this.prisma.user.delete({ where: { id: entity.id } });
+    return result !== null;
   }
 
-  updateAddress(user: UserEntity): Promise<void> {
+  async updateAddress(user: UserEntity): Promise<void> {
     const address = user.getPropsCopy().address;
 
-    const qb = this.em.createQueryBuilder(UserRecord, 'u');
-    return qb
-      .update({
+    const result = this.prisma.user.update({
+      data: {
+        postalCode: address.postalCode,
         country: address.country,
         street: address.street,
-        postalCode: address.postalCode,
-      })
-      .where({ id: user.id })
-      .execute();
+      },
+      where: {
+        id: user.id,
+      },
+    });
+    await Promise.resolve(result);
   }
 
   async findOneByEmail(email: string): Promise<UserEntity> {
-    const qb = this.em.createQueryBuilder(UserRecord, 'u');
-    const result = await qb.select('*').where({ email }).execute();
-    return result[0] ? this.mapper.toDomain(result[0]) : null;
+    const result = await this.prisma.user.findFirst({
+      where: {
+        email: email,
+      },
+    });
+    return result ? this.mapper.toDomain(result) : null;
   }
 }
