+++ backend/apps/sample-service/src/user/presentation/index.ts
@@ -1,3 +1,2 @@
-export * from '../application/dtos';
 export * from './user.http.controller';
 export * from './user.message.controller';
