+++ backend/apps/sample-service/src/user/user.mapper.ts
@@ -2,7 +2,6 @@ import { Mapper } from '@lib/ddd';
 import { UserEntity, AddressVO, UserRoles } from './domain';
 import { Injectable } from '@nestjs/common';
 import { UserResponseDto } from './presentation';
-import { UserRecord } from './infrastructure';
 import { User } from '@prisma/client/sample';
 
 /**
@@ -13,65 +12,7 @@ import { User } from '@prisma/client/sample';
  */
 
 @Injectable()
-export class UserMapper
-  implements Mapper<UserEntity, UserRecord, UserResponseDto>
-{
-  toPersistence(entity: UserEntity): UserRecord {
-    const copy = entity.getPropsCopy();
-    const record: UserRecord = {
-      id: copy.id,
-      createdAt: copy.createdAt,
-      updatedAt: copy.updatedAt,
-      email: copy.email,
-      country: copy.address.country,
-      postalCode: copy.address.postalCode,
-      street: copy.address.street,
-      role: copy.role,
-    };
-    return record;
-  }
-
-  toDomain(record: UserRecord): UserEntity {
-    const entity = new UserEntity({
-      id: record.id,
-      createdAt: new Date(record.createdAt),
-      updatedAt: new Date(record.updatedAt),
-      props: {
-        email: record.email,
-        role: record.role,
-        address: new AddressVO({
-          street: record.street,
-          postalCode: record.postalCode,
-          country: record.country,
-        }),
-      },
-    });
-    return entity;
-  }
-
-  toResponse(entity: UserEntity): UserResponseDto {
-    const props = entity.getPropsCopy();
-    const response = new UserResponseDto(entity);
-    response.email = props.email;
-    response.country = props.address.country;
-    response.postalCode = props.address.postalCode;
-    response.street = props.address.street;
-    return response;
-  }
-
-  /* ^ Data returned to the user is whitelisted to avoid leaks.
-     If a new property is added, like password or a
-     credit card number, it won't be returned
-     unless you specifically allow this.
-     (avoid blacklisting, which will return everything
-      but blacklisted items, which can lead to a data leak).
-  */
-}
-
-@Injectable()
-export class UserPrismaMapper
-  implements Mapper<UserEntity, User, UserResponseDto>
-{
+export class UserMapper implements Mapper<UserEntity, User, UserResponseDto> {
   toPersistence(entity: UserEntity): User {
     const copy = entity.getPropsCopy();
     const record: User = {
