+++ backend/apps/sample-service/src/user/user.module.ts
@@ -1,6 +1,5 @@
 import { Logger, Module, Provider } from '@nestjs/common';
-import { UserRecord, UserRepository } from './infrastructure';
-import { UserMapper, UserPrismaMapper } from './user.mapper';
+import { UserMapper } from './user.mapper';
 import { CqrsModule } from '@nestjs/cqrs';
 import { USER_REPOSITORY } from './user.di-tokens';
 import { UserHttpController, UserMessageController } from './presentation';
@@ -9,8 +8,7 @@ import {
   DeleteUserCommandHandler,
   FindUsersQueryHandler,
 } from './application';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
-import { UserPrismaRepository } from './infrastructure/user.prisma.repo';
+import { UserRepository } from './infrastructure';
 import { PrismaSampleService } from '@lib/sample-db';
 
 const httpControllers = [UserHttpController];
@@ -24,10 +22,10 @@ const commandHandlers: Provider[] = [
 
 const queryHandlers: Provider[] = [FindUsersQueryHandler];
 
-const mappers: Provider[] = [UserPrismaMapper];
+const mappers: Provider[] = [UserMapper];
 
 const repositories: Provider[] = [
-  { provide: USER_REPOSITORY, useClass: UserPrismaRepository },
+  { provide: USER_REPOSITORY, useClass: UserRepository },
 ];
 
 @Module({
