+++ backend/apps/sample-service/src/wallet/infrastructure/index.ts
@@ -1,2 +1,1 @@
-export * from './wallet.record';
 export * from './wallet.repository';
