+++ backend/package.json
@@ -5,9 +5,6 @@
   "scripts": {},
   "private": true,
   "dependencies": {
-    "@mikro-orm/cli": "^5.8.5",
-    "@mikro-orm/core": "^5.8.5",
-    "@mikro-orm/mysql": "^5.8.5",
     "@mikro-orm/nestjs": "^5.2.2",
     "@nestjs/common": "10.0.2",
     "@nestjs/core": "10.0.2",
